refactor: extract tooltip container to own component
