+++ projects/sbb-esta/angular-business/tooltip/src/public_api.ts
@@ -1,6 +1,7 @@
 export * from './tooltip.module';
 export * from './tooltip/tooltip.component';
 export * from './tooltip/tooltip';
+export * from './tooltip/tooltip-container.component';
 export * from './tooltip/tooltip-animations';
 export {
   TooltipBase,
