+++ projects/sbb-esta/angular-business/tooltip/src/tooltip/tooltip-container.component.html
@@ -1,0 +1,21 @@
+<div
+  class="sbb-tooltip sbb-tooltip-content"
+  [ngClass]="tooltipClass"
+  [class.sbb-tooltip-content-below]="
+    connectionPositionPair && connectionPositionPair.originY === 'bottom'
+  "
+  [class.sbb-tooltip-content-above]="
+    connectionPositionPair && connectionPositionPair.originY === 'top'
+  "
+  [class.sbb-tooltip-content-left]="
+    connectionPositionPair && connectionPositionPair.originX === 'end'
+  "
+  [class.sbb-tooltip-content-right]="
+    connectionPositionPair && connectionPositionPair.originX === 'start'
+  "
+  [@state]="_visibility"
+  (@state.start)="_animationStart()"
+  (@state.done)="_animationDone($event)"
+>
+  {{ message }}
+</div>
