+++ projects/sbb-esta/angular-business/tooltip/src/tooltip/tooltip-container.component.scss
